+++ air/src/constraints/stack/stack_manipulation/mod.rs
@@ -77,8 +77,10 @@ pub fn enforce_constraints<E: FieldElement>(
 // TRANSITION CONSTRAINT HELPERS
 // ================================================================================================
 
-/// Enforces constraints of the PAD operation. The PAD operation pushes a ZERO onto
-/// the stack. Therefore, the following constraints are enforced:
+/// Enforces constraints of the PAD operation.
+///
+/// The PAD operation pushes a ZERO onto the stack. Therefore, the following constraints are
+/// enforced:
 /// - The top element in the next frame should be ZERO. s0` = 0.
 pub fn enforce_pad_constraints<E: FieldElement>(
     frame: &EvaluationFrame<E>,
@@ -166,8 +168,10 @@ pub fn enforce_dup_movup_n_constraints<E: FieldElement>(
     13
 }
 
-/// Enforces constraints of the SWAP operation. The SWAP operation swaps the first
-/// two elements in the stack. Therefore, the following constraints are enforced:
+/// Enforces constraints of the SWAP operation.
+///
+/// The SWAP operation swaps the first two elements in the stack. Therefore, the following
+/// constraints are enforced:
 /// - The first element in the current frame should be equal to the second element in the next
 ///   frame.
 /// - The second element in the current frame should be equal to the first element in the next