+++ core/src/mast/serialization/info.rs
@@ -1,7 +1,7 @@
 use miden_crypto::hash::rpo::RpoDigest;
 use winter_utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
 
-use super::{basic_block_data_decoder::BasicBlockDataDecoder, DataOffset};
+use super::{basic_blocks::BasicBlockDataDecoder, NodeDataOffset};
 use crate::mast::{
     BasicBlockNode, CallNode, JoinNode, LoopNode, MastForest, MastNode, MastNodeId, SplitNode,
 };
@@ -21,24 +21,40 @@ pub struct MastNodeInfo {
 }
 
 impl MastNodeInfo {
-    pub fn new(mast_node: &MastNode, basic_block_offset: DataOffset) -> Self {
-        let ty = MastNodeType::new(mast_node, basic_block_offset);
+    /// Constructs a new [`MastNodeInfo`] from a [`MastNode`], along with an `ops_offset` and
+    /// `decorator_list_offset` in the case of [`BasicBlockNode`].
+    ///
+    /// If the represented [`MastNode`] is a [`BasicBlockNode`] that has an empty decorator list,
+    /// use `MastForest::MAX_DECORATORS` for the value of `decorator_list_offset`. For non-basic
+    /// block nodes, `ops_offset` and `decorator_list_offset` are ignored, and should be set to 0.
+    pub fn new(
+        mast_node: &MastNode,
+        ops_offset: NodeDataOffset,
+        decorator_list_offset: NodeDataOffset,
+    ) -> Self {
+        if !matches!(mast_node, &MastNode::Block(_)) {
+            debug_assert_eq!(ops_offset, 0);
+            debug_assert_eq!(decorator_list_offset, 0);
+        }
+
+        let ty = MastNodeType::new(mast_node, ops_offset, decorator_list_offset);
 
         Self { ty, digest: mast_node.digest() }
     }
 
     pub fn try_into_mast_node(
         self,
-        mast_forest: &MastForest,
+        mast_forest: &mut MastForest,
         basic_block_data_decoder: &BasicBlockDataDecoder,
     ) -> Result<MastNode, DeserializationError> {
         match self.ty {
-            MastNodeType::Block {
-                offset,
-                len: num_operations_and_decorators,
-            } => {
+            MastNodeType::Block { ops_offset, decorator_list_offset } => {
                 let (operations, decorators) = basic_block_data_decoder
-                    .decode_operations_and_decorators(offset, num_operations_and_decorators)?;
+                    .decode_operations_and_decorators(
+                        ops_offset,
+                        decorator_list_offset,
+                        mast_forest,
+                    )?;
                 let block = BasicBlockNode::new_unsafe(operations, decorators, self.digest);
                 Ok(MastNode::Block(block))
             },
@@ -126,10 +142,10 @@ pub enum MastNodeType {
         body_id: u32,
     } = LOOP,
     Block {
-        /// Offset of the basic block in the data segment
-        offset: u32,
-        /// The number of operations and decorators in the basic block
-        len: u32,
+        // offset of operations in node data
+        ops_offset: u32,
+        // offset of DecoratorList in node data
+        decorator_list_offset: u32,
     } = BLOCK,
     Call {
         callee_id: u32,
@@ -144,15 +160,18 @@ pub enum MastNodeType {
 /// Constructors
 impl MastNodeType {
     /// Constructs a new [`MastNodeType`] from a [`MastNode`].
-    pub fn new(mast_node: &MastNode, basic_block_offset: u32) -> Self {
+    ///
+    /// If the represented [`MastNode`] is a [`BasicBlockNode`] that has an empty decorator list,
+    /// use `MastForest::MAX_DECORATORS` for the value of `decorator_list_offset`.
+    pub fn new(
+        mast_node: &MastNode,
+        ops_offset: NodeDataOffset,
+        decorator_list_offset: NodeDataOffset,
+    ) -> Self {
         use MastNode::*;
 
         match mast_node {
-            Block(block_node) => {
-                let len = block_node.num_operations_and_decorators();
-
-                Self::Block { len, offset: basic_block_offset }
-            },
+            Block(_block_node) => Self::Block { decorator_list_offset, ops_offset },
             Join(join_node) => Self::Join {
                 left_child_id: join_node.first().0,
                 right_child_id: join_node.second().0,
@@ -169,7 +188,7 @@ impl MastNodeType {
                     Self::Call { callee_id: call_node.callee().0 }
                 }
             },
-            Dyn => Self::Dyn,
+            Dyn(_) => Self::Dyn,
             External(_) => Self::External,
         }
     }
@@ -190,7 +209,9 @@ impl Serializable for MastNodeType {
                 else_branch_id: else_branch,
             } => Self::encode_u32_pair(if_branch, else_branch),
             MastNodeType::Loop { body_id: body } => Self::encode_u32_payload(body),
-            MastNodeType::Block { offset, len } => Self::encode_u32_pair(offset, len),
+            MastNodeType::Block { ops_offset, decorator_list_offset } => {
+                Self::encode_u32_pair(ops_offset, decorator_list_offset)
+            },
             MastNodeType::Call { callee_id } => Self::encode_u32_payload(callee_id),
             MastNodeType::SysCall { callee_id } => Self::encode_u32_payload(callee_id),
             MastNodeType::Dyn => 0,
@@ -264,8 +285,8 @@ impl Deserializable for MastNodeType {
                 Ok(Self::Loop { body_id })
             },
             BLOCK => {
-                let (offset, len) = Self::decode_u32_pair(payload);
-                Ok(Self::Block { offset, len })
+                let (ops_offset, decorator_list_offset) = Self::decode_u32_pair(payload);
+                Ok(Self::Block { ops_offset, decorator_list_offset })
             },
             CALL => {
                 let callee_id = Self::decode_u32_payload(payload)?;