+++ CHANGELOG.md
@@ -9,10 +9,23 @@
 - Added `miden_core::utils::sync::racy_lock` module (#1463).
 - Updated `miden_core::utils` to re-export `std::sync::LazyLock` and `racy_lock::RacyLock as LazyLock` for std and no_std environments, respectively (#1463).
 - Made the undocumented behavior of the VM with regard to undefined behavior of u32 operations, stricter (#1480)
+- Introduced the `Emit` instruction (#1496)
 
 #### Fixes
 
 - Fixed an issue with formatting of blocks in Miden Assembly syntax
+
+#### Fixes
+
+- Decorators are now allowed in empty basic blocks (#1466)
+
+
+## 0.10.6 (2024-09-12) - `miden-processor` crate only.
+
+#### Enhancements
+
+- Added `PartialEq`, `Eq`, `Serialize` and `Deserialize` to `AdviceMap` and `AdviceInputs` structs (#1494).
+
 
 ## 0.10.5 (2024-08-21)